Move screenshot-related classes into own package
