+++ bugshaker/config/findbugs.xml
@@ -17,7 +17,7 @@ FindBugs custom exclude filter
 
     <!-- Custom excludes -->
     <Match>
-        <Class name="com.github.stkent.bugshaker.ScreenshotUriObservable" />
+        <Class name="com.github.stkent.bugshaker.screenshots.ScreenshotUriObservable" />
         <Method name="getScreenshotFile" />
         <Bug pattern="RV_RETURN_VALUE_IGNORED_BAD_PRACTICE" />
     </Match>
