+++ bugshaker/src/main/java/com/github/stkent/bugshaker/FeedbackEmailFlowManager.java
@@ -29,6 +29,8 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.view.WindowManager;
 
+import com.github.stkent.bugshaker.screenshots.ScreenshotProvider;
+
 import java.util.Arrays;
 import java.util.List;
