+++ bugshaker/src/main/java/com/github/stkent/bugshaker/screenshots/BaseScreenshotProvider.java
@@ -14,7 +14,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package com.github.stkent.bugshaker;
+package com.github.stkent.bugshaker.screenshots;
 
 import android.app.Activity;
 import android.content.Context;
