+++ bugshaker/src/main/java/com/github/stkent/bugshaker/screenshots/MapBitmapObservable.java
@@ -14,7 +14,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package com.github.stkent.bugshaker;
+package com.github.stkent.bugshaker.screenshots;
 
 import android.graphics.Bitmap;
 import android.support.annotation.NonNull;
